krapper_gen: separate the pure parse-output model from libclang-C cursor construction (front-end seam, #44 brick 1)
model/ and model/type/ now compile with ZERO clang.*/kotlinx.cinterop imports:
all construction-from-cursor code moved to two new front-end files in the
parsing layer (package com.monkopedia.krapper.generator):

- ModelFactories.kt: WrappedElement.Companion.mapAll/map (+ droppedUsrs/
  elementLookup state, isNestedInClassTemplate, isCopyAssignmentOf), cursor
  factory functions for WrappedClass/WrappedField/WrappedMethod/WrappedArgument/
  WrappedTemplate/WrappedTemplateParam/WrappedTypedef (former secondary
  constructors), and the iterator_range extraction (extractRangeReturn et al,
  formerly in model/RangeReturn.kt).
- TypeFactories.kt: WrappedType.Companion.invoke(CValue<CXType>, ...) extension
  operators + createForType and the typedef reducers (function-pointer/
  reference/pointer/assoc/size/c-alias), formerly the cursor half of the
  WrappedType companion.

The model keeps its API shape (same call syntax everywhere via factory
functions/companion extensions); WrappedType's companion keeps the pure
(String) -> WrappedType factory and the pointerTo/referenceTo/arrayOf/const
algebra. Dead code dropped with the move: WrappedClass.kt's unused fileParent
val and WrappedTemplateParam's unused companion determineType.

Pure refactor, zero behavior change intended.
